Add hint icon to in-browser UI community projects (#679)
This PR contains a few commits, but the punchline is that we add a "🔎"
icon next to community projects that provide an in-browser UI for the
registry. The intention is that these projects can serve as a stopgap or
substitute for
https://github.com/modelcontextprotocol/registry/issues/99.

This is just a first take.  Very open to feedback!